move resources into external files
